+++ src/Adapter/Logger.php
@@ -9,7 +9,10 @@ use Innmind\Filesystem\{
     Name,
     Directory,
 };
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Attempt,
+};
 use Psr\Log\LoggerInterface;
 
 final class Logger implements Adapter
@@ -29,10 +32,11 @@ final class Logger implements Adapter
     }
 
     #[\Override]
-    public function add(File|Directory $file): void
+    public function add(File|Directory $file): Attempt
     {
         $this->logger->debug('Adding file {file}', ['file' => $file->name()->toString()]);
-        $this->filesystem->add($file);
+
+        return $this->filesystem->add($file);
     }
 
     #[\Override]
@@ -64,10 +68,11 @@ final class Logger implements Adapter
     }
 
     #[\Override]
-    public function remove(Name $file): void
+    public function remove(Name $file): Attempt
     {
         $this->logger->debug('Removing file {file}', ['file' => $file->toString()]);
-        $this->filesystem->remove($file);
+
+        return $this->filesystem->remove($file);
     }
 
     #[\Override]